# Decompose linalg.conv_2d_q to linalg.conv_2d