--- a/compiler/src/iree/compiler/InputConversion/Common/CMakeLists.txt
+++ b/compiler/src/iree/compiler/InputConversion/Common/CMakeLists.txt
@@ -42,6 +42,7 @@ iree_cc_library(
     "IREEImportPublic.cpp"
     "ImportMLProgram.cpp"
     "Passes.cpp"
+    "QuantizedConvToConv.cpp"
     "QuantizedMatmulToMatmul.cpp"
     "SanitizeModuleNames.cpp"
     "TopLevelSCFToCFG.cpp"