--- a/compiler/src/iree/compiler/InputConversion/Common/CMakeLists.txt
+++ b/compiler/src/iree/compiler/InputConversion/Common/CMakeLists.txt
@@ -46,6 +46,7 @@ iree_cc_library(
     "QuantizedMatmulToMatmul.cpp"
     "SanitizeModuleNames.cpp"
     "TopLevelSCFToCFG.cpp"
+    "Utils.cpp"
   DEPS
     ::PassHeaders
     ::PassesIncGen