--- a/compiler/src/iree/compiler/InputConversion/Common/Passes.td
+++ b/compiler/src/iree/compiler/InputConversion/Common/Passes.td
@@ -21,6 +21,13 @@ def ImportMLProgram :
   let constructor = "mlir::iree_compiler::createImportMLProgramPass()";
 }
 
+def LinalgQuantizedConvToConvPass
+    : Pass<"iree-linalg-quantized-conv-to-conv", "func::FuncOp"> {
+  let summary = "lower quantized_conv to conv";
+  let constructor = "mlir::iree_compiler::createLinalgQuantizedConvToConvPass()";
+  // let dependentDialects = ["mlir::linalg::LinalgDialect"];
+}
+
 def LinalgQuantizedMatmulToMatmulPass
     : Pass<"iree-linalg-quantized-matmul-to-matmul", "func::FuncOp"> {
   let summary = "lower quantized_matmul to matmul";