--- a/compiler/src/iree/compiler/InputConversion/Common/test/BUILD
+++ b/compiler/src/iree/compiler/InputConversion/Common/test/BUILD
@@ -20,6 +20,7 @@ iree_lit_test_suite(
         [
             "import_ml_program.mlir",
             "iree_import_public.mlir",
+            "linalg_quantized_conv_to_conv.mlir",
             "linalg_quantized_matmul_to_matmul.mlir",
             "sanitize_module_names.mlir",
             "top_level_scf_to_cfg.mlir",