linalg.conv_2d_nhwc_hwcf_q includes the zeropoint modification within the inner loop. This can be decomposed to series of generic and pooling computations. This allows the tiling/vectorization code to run on the quantized variants and decreases the total additions / sums.