@rsuderman 2022-10-14
Previous performance:
```
Running tools/iree-benchmark-module
Run on (96 X 2000.15 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x48)
  L1 Instruction 32 KiB (x48)
  L2 Unified 1024 KiB (x48)
  L3 Unified 39424 KiB (x2)
Load Average: 0.30, 0.42, 3.45
-----------------------------------------------------------------------------------------
Benchmark                               Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------------------------
BM_main/process_time/real_time       7.48 ms         58.1 ms           89 items_per_second=133.607/s
```

New Performance:
```
2022-10-14T05:27:28+00:00
Running tools/iree-benchmark-module
Run on (96 X 2000.15 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x48)
  L1 Instruction 32 KiB (x48)
  L2 Unified 1024 KiB (x48)
  L3 Unified 39424 KiB (x2)
Load Average: 0.28, 0.41, 3.43
-----------------------------------------------------------------------------------------
Benchmark                               Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------------------------
BM_main/process_time/real_time       5.32 ms         39.5 ms          129 items_per_second=187.856/s
```